+++ tests/test_climate.py
@@ -130,8 +130,8 @@ async def test_cerfthermostat(
     assert climate.hvac_mode == HmHvacMode.AUTO
     assert climate.hvac_modes == [HmHvacMode.AUTO, HmHvacMode.HEAT, HmHvacMode.OFF]
     await climate.set_hvac_mode(HmHvacMode.HEAT)
-    assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU0000050:4", paramset_key="VALUES", value={"MANU_MODE": 12.0}
+    assert mock_client.method_calls[-1] == call.set_value(
+        channel_address="VCU0000050:4", paramset_key="VALUES", parameter="MANU_MODE", value=12.0
     )
     central.event(const.LOCAL_INTERFACE_ID, "VCU0000050:4", "CONTROL_MODE", HMIP_MODE_MANU)
     assert climate.hvac_mode == HmHvacMode.HEAT
@@ -147,8 +147,8 @@ async def test_cerfthermostat(
     assert climate.hvac_action == HmHvacAction.OFF
 
     await climate.set_hvac_mode(HmHvacMode.AUTO)
-    assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU0000050:4", paramset_key="VALUES", value={"AUTO_MODE": True}
+    assert mock_client.method_calls[-1] == call.set_value(
+        channel_address="VCU0000050:4", paramset_key="VALUES", parameter="AUTO_MODE", value=True
     )
     central.event(const.LOCAL_INTERFACE_ID, "VCU0000050:4", "CONTROL_MODE", 0)
     central.event(const.LOCAL_INTERFACE_ID, "VCU0000050:4", "SET_TEMPERATURE", 24.0)
@@ -254,8 +254,8 @@ async def test_ceipthermostat(
         HmPresetMode.NONE,
     ]
     await climate.set_preset_mode(HmPresetMode.BOOST)
-    assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1", paramset_key="VALUES", value={"BOOST_MODE": True}
+    assert mock_client.method_calls[-1] == call.set_value(
+        channel_address="VCU1769958:1", paramset_key="VALUES", parameter="BOOST_MODE", value=True
     )
     central.event(const.LOCAL_INTERFACE_ID, "VCU1769958:1", "BOOST_MODE", 1)
     assert climate.preset_mode == HmPresetMode.BOOST
@@ -280,16 +280,16 @@ async def test_ceipthermostat(
         "week_program_6",
     ]
     await climate.set_preset_mode(HmPresetMode.NONE)
-    assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1", paramset_key="VALUES", value={"BOOST_MODE": False}
+    assert mock_client.method_calls[-1] == call.set_value(
+        channel_address="VCU1769958:1", paramset_key="VALUES", parameter="BOOST_MODE", value=False
     )
     central.event(const.LOCAL_INTERFACE_ID, "VCU1769958:1", "SET_POINT_MODE", HMIP_MODE_AWAY)
     assert climate.preset_mode == HmPresetMode.AWAY
 
     central.event(const.LOCAL_INTERFACE_ID, "VCU1769958:1", "SET_POINT_MODE", HMIP_MODE_AUTO)
     await climate.set_preset_mode(HmPresetMode.WEEK_PROGRAM_1)
-    assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1", paramset_key="VALUES", value={"ACTIVE_PROFILE": 1}
+    assert mock_client.method_calls[-1] == call.set_value(
+        channel_address="VCU1769958:1", paramset_key="VALUES", parameter="ACTIVE_PROFILE", value=1
     )
     assert climate.preset_mode == HmPresetMode.WEEK_PROGRAM_1
 
@@ -304,9 +304,7 @@ async def test_ceipthermostat(
     #     },
     # )
     assert mock_client.method_calls[-1] == call.put_paramset(
-        address="VCU1769958:1",
-        paramset_key="VALUES",
-        value={"SET_POINT_TEMPERATURE": 17.0},
+        address="VCU1769958:1", paramset_key="VALUES", value={"SET_POINT_TEMPERATURE": 17.0}
     )
     await climate.enable_away_mode_by_calendar(
         start=datetime(2000, 12, 1), end=datetime(2024, 12, 1), away_temperature=17.0