+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2023.2.2 (2023-02-01)
 - Add comments to parameter_visibility
+- Use put_paramset only if there are more than one parameters to send
 
 # Version 2023.2.1 (2023-02-01)
 - Separate check for parameter is un_ignored based on if it should be hidden or not